+++ packages/jds/src/components/Listbox/index.ts
@@ -1,0 +1,10 @@
+export { Listbox } from "./Listbox";
+
+export type {
+  ListboxProps,
+  SelectBaseProps,
+  SelectOption,
+  SelectionMode,
+  OptionVariant,
+  SelectDimension,
+} from "./listbox.types";
