+++ packages/jds/src/components/Listbox/listbox.utils.ts
