+++ packages/jds/src/components/MultiSelect/index.ts
@@ -1,0 +1,2 @@
+export { MultiSelect } from "./MultiSelect";
+export type { MultiSelectProps } from "./multiSelect.types";
