+++ packages/jds/src/components/MultiSelect/multiSelect.types.ts
@@ -1,0 +1,16 @@
+import type { SelectBaseProps } from "../Listbox";
+
+type MultiSelectControlledProps = {
+  value: string[];
+  defaultValue?: never;
+  onChange: (value: string[]) => void;
+};
+
+type MultiSelectUncontrolledProps = {
+  value?: never;
+  defaultValue?: string[];
+  onChange?: (value: string[]) => void;
+};
+
+export type MultiSelectProps = SelectBaseProps &
+  (MultiSelectControlledProps | MultiSelectUncontrolledProps);
